commit 1
I’ve tweaked this a bit but to make it more user friendly. But I think
the copy needs to be reordered  and broken down further info sections,
e.g. it should probably go straight into why there’s a need for Sistah
Space? Or perhaps the sections could all be links with drop down copy -
it just seems there is a lot on the page! Are you planning on adding an
image in the white space? I can’t remember what you said now.